requirements.txt
fastapi~=0.68.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart==0.0.20
# aiofiles~=24.1.0
# SQLAlchemy~=2.0.36
starlette~=0.14.2
PyYAML~=6.0.2
loguru~=0.7.3